Fix errors due to duplicate function ...
