+++ src/connectors/sso.ts
@@ -1,4 +1,6 @@
-﻿// tslint:disable-next-line
+﻿import { getQsParam } from './common';
+
+// tslint:disable-next-line
 require('./sso.scss');
 
 document.addEventListener('DOMContentLoaded', (event) => {
@@ -6,19 +8,3 @@ document.addEventListener('DOMContentLoaded', (event) => {
     const state = getQsParam('state');
     window.location.href = window.location.origin + '/#/sso?code=' + code + '&state=' + state;
 });
-
-function getQsParam(name: string) {
-    const url = window.location.href;
-    name = name.replace(/[\[\]]/g, '\\$&');
-    const regex = new RegExp('[?&]' + name + '(=([^&#]*)|&|#|$)');
-    const results = regex.exec(url);
-
-    if (!results) {
-        return null;
-    }
-    if (!results[2]) {
-        return '';
-    }
-
-    return decodeURIComponent(results[2].replace(/\+/g, ' '));
-}
