+++ src/connectors/webauthn.ts
@@ -1,3 +1,5 @@
+import { getQsParam } from './common';
+
 document.addEventListener('DOMContentLoaded', (event) => {
     init();
 });
@@ -128,22 +130,6 @@ function success(assertedCredential: PublicKeyCredential) {
 
 function info(message: string) {
     parent.postMessage('info|' + message, parentUrl);
-}
-
-function getQsParam(name: string) {
-    const url = window.location.href;
-    name = name.replace(/[\[\]]/g, '\\$&');
-    const regex = new RegExp('[?&]' + name + '(=([^&#]*)|&|#|$)');
-    const results = regex.exec(url);
-
-    if (!results) {
-        return null;
-    }
-    if (!results[2]) {
-        return '';
-    }
-
-    return decodeURIComponent(results[2].replace(/\+/g, ' '));
 }
 
 function b64Decode(str: string) {
